Move fromAirTableObject methods from Airtable data objects directly to Datasources modules
