+++ api/lib/infrastructure/datasources/airtable/area-datasource.js
@@ -10,18 +10,28 @@ const USED_FIELDS = [
   'Competences (identifiants)',
 ];
 
+function fromAirTableObject(airtableDomaineObject) {
+  return new Area({
+    id: airtableDomaineObject.getId(),
+    code: airtableDomaineObject.get('Code'),
+    name: airtableDomaineObject.get('Nom'),
+    title: airtableDomaineObject.get('Titre'),
+    competenceIds: airtableDomaineObject.get('Competences (identifiants)'),
+  });
+}
+
 module.exports = {
 
   tableName: TABLE_NAME,
 
   usedFields: USED_FIELDS,
 
+  fromAirTableObject,
+
   list() {
     return airtable.findRecords(TABLE_NAME, USED_FIELDS)
       .then((airtableRawObjects) => {
-        return airtableRawObjects.map((airtableRawObject) => {
-          return Area.fromAirTableObject(airtableRawObject);
-        });
+        return airtableRawObjects.map(fromAirTableObject);
       });
   },
 };
