+++ api/lib/infrastructure/datasources/airtable/objects/Area.js
@@ -12,16 +12,6 @@ class Area {
     this.title = title;
     this.competenceIds = competenceIds;
   }
-
-  static fromAirTableObject(airtableDomaineObject) {
-    return new Area({
-      id: airtableDomaineObject.getId(),
-      code: airtableDomaineObject.get('Code'),
-      name: airtableDomaineObject.get('Nom'),
-      title: airtableDomaineObject.get('Titre'),
-      competenceIds: airtableDomaineObject.get('Competences (identifiants)'),
-    });
-  }
 }
 
 module.exports = Area;
