+++ api/lib/infrastructure/datasources/airtable/objects/Challenge.js
@@ -42,52 +42,6 @@ class Challenge {
     this.embedHeight = embedHeight;
     this.illustrationAlt = illustrationAlt;
   }
-
-  static fromAirTableObject(airtableEpreuveObject) {
-
-    let illustrationUrl;
-    if (airtableEpreuveObject.get('Illustration de la consigne')) {
-      illustrationUrl = airtableEpreuveObject.get('Illustration de la consigne')[0].url;
-    }
-
-    let attachments;
-    if (airtableEpreuveObject.get('Pièce jointe')) {
-      attachments = airtableEpreuveObject.get('Pièce jointe').map((attachment) => attachment.url).reverse();
-    }
-
-    let competenceId;
-    if (airtableEpreuveObject.get('Compétences (via tube)')) {
-      competenceId = airtableEpreuveObject.get('Compétences (via tube)')[0];
-    }
-
-    let timer;
-    if (airtableEpreuveObject.get('Timer')) {
-      timer = parseInt(airtableEpreuveObject.get('Timer'));
-    }
-
-    return new Challenge({
-      id: airtableEpreuveObject.getId(),
-      instruction: airtableEpreuveObject.get('Consigne'),
-      proposals: airtableEpreuveObject.get('Propositions'),
-      type: airtableEpreuveObject.get('Type d\'épreuve'),
-      solution: airtableEpreuveObject.get('Bonnes réponses'),
-      t1Status: airtableEpreuveObject.get('T1 - Espaces, casse & accents'),
-      t2Status: airtableEpreuveObject.get('T2 - Ponctuation'),
-      t3Status: airtableEpreuveObject.get('T3 - Distance d\'édition'),
-      scoring: airtableEpreuveObject.get('Scoring'),
-      status: airtableEpreuveObject.get('Statut'),
-      skillIds: airtableEpreuveObject.get('Acquix'),
-      skills: airtableEpreuveObject.get('acquis'),
-      embedUrl: airtableEpreuveObject.get('Embed URL'),
-      embedTitle: airtableEpreuveObject.get('Embed title'),
-      embedHeight: airtableEpreuveObject.get('Embed height'),
-      timer,
-      illustrationUrl,
-      attachments,
-      competenceId,
-      illustrationAlt: airtableEpreuveObject.get('Texte alternatif illustration'),
-    });
-  }
 }
 
 module.exports = Challenge;
