+++ api/lib/infrastructure/datasources/airtable/objects/Competence.js
@@ -16,19 +16,6 @@ class Competence {
     this.skillIds = skillIds;
     this.areaId = areaId;
   }
-
-  static fromAirTableObject(rawAirtableCompetence) {
-
-    return new Competence({
-      id: rawAirtableCompetence.getId(),
-      name: rawAirtableCompetence.get('Titre'),
-      index: rawAirtableCompetence.get('Sous-domaine'),
-      description: rawAirtableCompetence.get('Description'),
-      areaId: rawAirtableCompetence.get('Domaine') ? rawAirtableCompetence.get('Domaine')[0] : '',
-      courseId: rawAirtableCompetence.get('Tests') ? rawAirtableCompetence.get('Tests')[0] : '',
-      skillIds: rawAirtableCompetence.get('Acquis (via Tubes)'),
-    });
-  }
 }
 
 module.exports = Competence;
