+++ api/lib/infrastructure/datasources/airtable/objects/Course.js
@@ -17,22 +17,6 @@ class Course {
     this.challenges = challenges;
   }
 
-  static fromAirTableObject(airtableRecord) {
-    let imageUrl;
-    if (airtableRecord.get('Image')) {
-      imageUrl = airtableRecord.get('Image')[0].url;
-    }
-
-    return new Course({
-      id: airtableRecord.getId(),
-      name: airtableRecord.get('Nom'),
-      description: airtableRecord.get('Description'),
-      adaptive: airtableRecord.get('Adaptatif ?'),
-      competences: airtableRecord.get('Competence'),
-      challenges: airtableRecord.get('Épreuves'),
-      imageUrl,
-    });
-  }
 }
 
 module.exports = Course;
