+++ api/lib/infrastructure/datasources/airtable/objects/Skill.js
@@ -18,19 +18,6 @@ class Skill {
     this.pixValue = pixValue;
     this.competenceId = competenceId;
   }
-
-  static fromAirTableObject(airtableSkillObject) {
-    return new Skill({
-      id: airtableSkillObject.getId(),
-      name: airtableSkillObject.get('Nom'),
-      hint: airtableSkillObject.get('Indice'),
-      hintStatus: airtableSkillObject.get('Statut de l\'indice'),
-      tutorialIds: airtableSkillObject.get('Comprendre'),
-      learningMoreTutorialIds: airtableSkillObject.get('En savoir plus'),
-      pixValue: airtableSkillObject.get('PixValue'),
-      competenceId: airtableSkillObject.get('Compétence (via Tube)')[0],
-    });
-  }
 }
 
 module.exports = Skill;
