+++ api/lib/infrastructure/datasources/airtable/objects/Tutorial.js
@@ -14,17 +14,6 @@ class Tutorial {
     this.source = source;
     this.title = title;
   }
-
-  static fromAirTableObject(airtableTutorialObject) {
-    return new Tutorial({
-      id: airtableTutorialObject.getId(),
-      duration: airtableTutorialObject.get('Durée'),
-      format: airtableTutorialObject.get('Format'),
-      link: airtableTutorialObject.get('Lien'),
-      source: airtableTutorialObject.get('Source'),
-      title: airtableTutorialObject.get('Titre'),
-    });
-  }
 }
 
 module.exports = Tutorial;
