+++ api/lib/infrastructure/datasources/airtable/tutorial-datasource.js
@@ -13,12 +13,23 @@ const USED_FIELDS = [
   'Titre',
 ];
 
+function fromAirTableObject(airtableTutorialObject) {
+  return new Tutorial({
+    id: airtableTutorialObject.getId(),
+    duration: airtableTutorialObject.get('Durée'),
+    format: airtableTutorialObject.get('Format'),
+    link: airtableTutorialObject.get('Lien'),
+    source: airtableTutorialObject.get('Source'),
+    title: airtableTutorialObject.get('Titre'),
+  });
+}
+
 function _doQuery(filter) {
   return airtable.findRecords(TABLE_NAME, USED_FIELDS)
     .then((rawTutorials) => {
       return _(rawTutorials)
         .filter(filter)
-        .map(Tutorial.fromAirTableObject)
+        .map(fromAirTableObject)
         .value();
     });
 }
@@ -29,13 +40,15 @@ module.exports = {
 
   usedFields: USED_FIELDS,
 
+  fromAirTableObject,
+
   findByRecordIds(tutorialRecordIds) {
     return _doQuery((rawTutorial) => _.includes(tutorialRecordIds, rawTutorial.id));
   },
 
   get(id) {
     return airtable.getRecord(TABLE_NAME, id)
-      .then(Tutorial.fromAirTableObject)
+      .then(fromAirTableObject)
       .catch((err) => {
         if (err.error === 'NOT_FOUND') {
           throw new AirtableResourceNotFound();
