+++ api/tests/unit/infrastructure/datasources/airtable/area-datasource_test.js
@@ -1,10 +1,25 @@
-const { expect, sinon } = require('../../../../test-helper');
+const { expect, sinon, domainBuilder } = require('../../../../test-helper');
 const airtable = require('../../../../../lib/infrastructure/airtable');
 const areaDatasource = require('../../../../../lib/infrastructure/datasources/airtable/area-datasource');
 const areaRawAirTableFixture = require('../../../../tooling/fixtures/infrastructure/areaRawAirTableFixture');
 const { Area } = require('../../../../../lib/infrastructure/datasources/airtable/objects');
 
 describe('Unit | Infrastructure | Datasource | Airtable | AreaDatasource', () => {
+
+  describe('#fromAirTableObject', () => {
+
+    it('should create a Area from the AirtableRecord', () => {
+      // given
+      const expectedArea = domainBuilder.buildAreaAirtableDataObject();
+
+      // when
+      const area = areaDatasource.fromAirTableObject(areaRawAirTableFixture());
+
+      // then
+      expect(area).to.be.an.instanceof(Area);
+      expect(area).to.deep.equal(expectedArea);
+    });
+  });
 
   describe('#list', () => {
