+++ api/tests/unit/infrastructure/datasources/airtable/challenge-datasource_test.js
@@ -2,6 +2,7 @@ const { expect, sinon } = require('../../../../test-helper');
 const airtable = require('../../../../../lib/infrastructure/airtable');
 const AirtableError = require('airtable').Error;
 const challengeDatasource = require('../../../../../lib/infrastructure/datasources/airtable/challenge-datasource');
+const challengeAirtableDataObjectFixture = require('../../../../tooling/fixtures/infrastructure/challengeAirtableDataObjectFixture');
 const challengeRawAirTableFixture = require('../../../../tooling/fixtures/infrastructure/challengeRawAirTableFixture');
 const { Challenge } = require('../../../../../lib/infrastructure/datasources/airtable/objects');
 const AirtableResourceNotFound = require('../../../../../lib/infrastructure/datasources/airtable/AirtableResourceNotFound');
@@ -180,4 +181,73 @@ describe('Unit | Infrastructure | Datasource | Airtable | ChallengeDatasource', 
       });
     });
   });
+
+  describe('#fromAirTableObject', () => {
+
+    it('should create a Challenge from the AirtableRecord', () => {
+      // given
+      const expectedChallenge = challengeAirtableDataObjectFixture();
+
+      // when
+      const challenge = challengeDatasource.fromAirTableObject(challengeRawAirTableFixture());
+
+      // then
+      expect(challenge).to.be.an.instanceof(Challenge);
+      expect(challenge).to.deep.equal(expectedChallenge);
+    });
+
+    it('should deal with a missing illustration', () => {
+      // given
+      const airtableEpreuveObject = challengeRawAirTableFixture();
+      airtableEpreuveObject.set('Illustration de la consigne', undefined);
+
+      // when
+      const challenge = challengeDatasource.fromAirTableObject(airtableEpreuveObject);
+
+      // then
+      expect(challenge).to.be.an.instanceof(Challenge);
+      expect(challenge.illustrationUrl).to.be.undefined;
+    });
+
+    it('should deal with a missing timer', () => {
+      // given
+      const airtableEpreuveObject = challengeRawAirTableFixture();
+      airtableEpreuveObject.set('Timer', undefined);
+
+      // when
+      const challenge = challengeDatasource.fromAirTableObject(airtableEpreuveObject);
+
+      // then
+      expect(challenge).to.be.an.instanceof(Challenge);
+      expect(challenge.timer).to.be.undefined;
+    });
+
+    it('should deal with a missing Pièce jointe', () => {
+      // given
+      const airtableEpreuveObject = challengeRawAirTableFixture();
+      airtableEpreuveObject.set('Pièce jointe', undefined);
+
+      // when
+      const challenge = challengeDatasource.fromAirTableObject(airtableEpreuveObject);
+
+      // then
+      expect(challenge).to.be.an.instanceof(Challenge);
+      expect(challenge.attachments).to.be.undefined;
+    });
+
+    it('should deal with a missing competences', () => {
+      // given
+      const airtableEpreuveObject = challengeRawAirTableFixture();
+      airtableEpreuveObject.set('competences', undefined);
+      airtableEpreuveObject.set('Compétences (via tube)', undefined);
+
+      // when
+      const challenge = challengeDatasource.fromAirTableObject(airtableEpreuveObject);
+
+      // then
+      expect(challenge).to.be.an.instanceof(Challenge);
+      expect(challenge.competenceId).to.be.undefined;
+    });
+  });
+
 });
