+++ api/tests/unit/infrastructure/datasources/airtable/competence-datasource_test.js
@@ -1,10 +1,25 @@
-const { expect, sinon } = require('../../../../test-helper');
+const { expect, sinon, domainBuilder } = require('../../../../test-helper');
 const airtable = require('../../../../../lib/infrastructure/airtable');
 const competenceDatasource = require('../../../../../lib/infrastructure/datasources/airtable/competence-datasource');
 const competenceRawAirTableFixture = require('../../../../tooling/fixtures/infrastructure/competenceRawAirTableFixture');
-const airTableDataModels = require('../../../../../lib/infrastructure/datasources/airtable/objects');
+const { Competence } = require('../../../../../lib/infrastructure/datasources/airtable/objects');
 
 describe('Unit | Infrastructure | Datasource | Airtable | CompetenceDatasource', () => {
+
+  describe('#fromAirTableObject', () => {
+
+    it('should create a Competence from the AirtableRecord', () => {
+      // given
+      const expectedCompetence = domainBuilder.buildCompetenceAirtableDataObject();
+
+      // when
+      const area = competenceDatasource.fromAirTableObject(competenceRawAirTableFixture());
+
+      // then
+      expect(area).to.be.an.instanceof(Competence);
+      expect(area).to.deep.equal(expectedCompetence);
+    });
+  });
 
   describe('#get', () => {
 
@@ -18,7 +33,7 @@ describe('Unit | Infrastructure | Datasource | Airtable | CompetenceDatasource',
       const competence = await competenceDatasource.get('recsvLz0W2ShyfD63');
 
       // then
-      expect(competence).to.be.an.instanceof(airTableDataModels.Competence);
+      expect(competence).to.be.an.instanceof(Competence);
       expect(competence.id).to.equal('recsvLz0W2ShyfD63');
       expect(competence.name).to.equal('Mener une recherche et une veille d’information');
     });
@@ -36,7 +51,7 @@ describe('Unit | Infrastructure | Datasource | Airtable | CompetenceDatasource',
       const competences = await competenceDatasource.list();
 
       // then
-      expect(competences[0]).to.be.an.instanceof(airTableDataModels.Competence);
+      expect(competences[0]).to.be.an.instanceof(Competence);
       expect(competences[0].id).to.equal('recsvLz0W2ShyfD63');
       expect(competences[0].name).to.equal('Mener une recherche et une veille d’information');
     });
