+++ api/tests/unit/infrastructure/datasources/airtable/models/Challenge_test.js
@@ -45,72 +45,4 @@ describe('Unit | Infrastructure | Datasource | Airtable | Model | Challenge', ()
       expect(challengeDataObject).to.deep.equal(challengeRawData);
     });
   });
-
-  describe('#fromAirTableObject', () => {
-
-    it('should create a Challenge from the AirtableRecord', () => {
-      // given
-      const expectedChallenge = ChallengeAirtableDataModelFixture();
-
-      // when
-      const challenge = dataModels.Challenge.fromAirTableObject(challengeRawAirTableFixture());
-
-      // then
-      expect(challenge).to.be.an.instanceof(dataModels.Challenge);
-      expect(challenge).to.deep.equal(expectedChallenge);
-    });
-
-    it('should deal with a missing illustration', () => {
-      // given
-      const airtableEpreuveObject = challengeRawAirTableFixture();
-      airtableEpreuveObject.set('Illustration de la consigne', undefined);
-
-      // when
-      const challenge = dataModels.Challenge.fromAirTableObject(airtableEpreuveObject);
-
-      // then
-      expect(challenge).to.be.an.instanceof(dataModels.Challenge);
-      expect(challenge.illustrationUrl).to.be.undefined;
-    });
-
-    it('should deal with a missing timer', () => {
-      // given
-      const airtableEpreuveObject = challengeRawAirTableFixture();
-      airtableEpreuveObject.set('Timer', undefined);
-
-      // when
-      const challenge = dataModels.Challenge.fromAirTableObject(airtableEpreuveObject);
-
-      // then
-      expect(challenge).to.be.an.instanceof(dataModels.Challenge);
-      expect(challenge.timer).to.be.undefined;
-    });
-
-    it('should deal with a missing Pièce jointe', () => {
-      // given
-      const airtableEpreuveObject = challengeRawAirTableFixture();
-      airtableEpreuveObject.set('Pièce jointe', undefined);
-
-      // when
-      const challenge = dataModels.Challenge.fromAirTableObject(airtableEpreuveObject);
-
-      // then
-      expect(challenge).to.be.an.instanceof(dataModels.Challenge);
-      expect(challenge.attachments).to.be.undefined;
-    });
-
-    it('should deal with a missing competences', () => {
-      // given
-      const airtableEpreuveObject = challengeRawAirTableFixture();
-      airtableEpreuveObject.set('competences', undefined);
-      airtableEpreuveObject.set('Compétences (via tube)', undefined);
-
-      // when
-      const challenge = dataModels.Challenge.fromAirTableObject(airtableEpreuveObject);
-
-      // then
-      expect(challenge).to.be.an.instanceof(dataModels.Challenge);
-      expect(challenge.competenceId).to.be.undefined;
-    });
-  });
 });
