--- api/tests/unit/infrastructure/datasources/airtable/models/Tutorial_test.js
@@ -1,22 +1,0 @@
-const dataModels = require('../../../../../../lib/infrastructure/datasources/airtable/objects/index');
-const tutorialRawAirTableFixture = require('../../../../../tooling/fixtures/infrastructure/tutorialRawAirtableFixture');
-const tutorialAirtableDataModelFixture = require('../../../../../tooling/fixtures/infrastructure/tutorialAirtableDataObjectFixture');
-const { expect } = require('../../../../../test-helper');
-
-describe('Unit | Infrastructure | Datasource | Airtable | Model | Tutorial', () => {
-
-  context('#fromAirTableObject', () => {
-
-    it('should create a Tutorial from the AirtableRecord', () => {
-      // given
-      const expectedTuto = tutorialAirtableDataModelFixture();
-
-      // when
-      const tuto = dataModels.Tutorial.fromAirTableObject(tutorialRawAirTableFixture());
-
-      // then
-      expect(tuto).to.be.an.instanceof(dataModels.Tutorial);
-      expect(tuto).to.deep.equal(expectedTuto);
-    });
-  });
-});
