+++ api/tests/unit/infrastructure/datasources/airtable/skill-datasource_test.js
@@ -1,6 +1,7 @@
 const { expect, sinon } = require('../../../../test-helper');
 const airtable = require('../../../../../lib/infrastructure/airtable');
 const skillDatasource = require('../../../../../lib/infrastructure/datasources/airtable/skill-datasource');
+const skillAirtableDataObjectFixture = require('../../../../tooling/fixtures/infrastructure/skillAirtableDataObjectFixture');
 const skillRawAirTableFixture = require('../../../../tooling/fixtures/infrastructure/skillRawAirTableFixture');
 const { Skill } = require('../../../../../lib/infrastructure/datasources/airtable/objects');
 const AirtableRecord = require('airtable').Record;
@@ -17,6 +18,21 @@ function makeAirtableFake(records) {
 }
 
 describe('Unit | Infrastructure | Datasource | Airtable | SkillDatasource', () => {
+
+  describe('#fromAirTableObject', () => {
+
+    it('should create a Skill from the AirtableRecord', () => {
+      // given
+      const expectedSkill = skillAirtableDataObjectFixture();
+
+      // when
+      const skill = skillDatasource.fromAirTableObject(skillRawAirTableFixture());
+
+      // then
+      expect(skill).to.be.an.instanceof(Skill);
+      expect(skill).to.deep.equal(expectedSkill);
+    });
+  });
 
   describe('#get', () => {
