+++ api/tests/unit/infrastructure/datasources/airtable/tutorial-datasource_test.js
@@ -2,9 +2,9 @@ const { expect, sinon } = require('../../../../test-helper');
 const airtable = require('../../../../../lib/infrastructure/airtable');
 const airTableDataModels = require('../../../../../lib/infrastructure/datasources/airtable/objects');
 const tutorialDatasource = require('../../../../../lib/infrastructure/datasources/airtable/tutorial-datasource');
+const tutorialAirtableDataObjectFixture = require('../../../../tooling/fixtures/infrastructure/tutorialAirtableDataObjectFixture');
 const tutorialRawAirTableFixture = require('../../../../tooling/fixtures/infrastructure/tutorialRawAirtableFixture');
-const AirtableRecord = require('airtable').Record;
-const AirtableError = require('airtable').Error;
+const { Record: AirtableRecord, Error: AirtableError } = require('airtable');
 const AirtableResourceNotFound = require('../../../../../lib/infrastructure/datasources/airtable/AirtableResourceNotFound');
 const { Tutorial } = require('../../../../../lib/infrastructure/datasources/airtable/objects');
 const _ = require('lodash');
@@ -21,18 +21,33 @@ function makeAirtableFake(records) {
 
 describe('Unit | Infrastructure | Datasource | Airtable | TutorialDatasource', () => {
 
+  describe('#fromAirTableObject', () => {
+
+    it('should create a Tutorial from the AirtableRecord', () => {
+      // given
+      const expectedTuto = tutorialAirtableDataObjectFixture();
+
+      // when
+      const tuto = tutorialDatasource.fromAirTableObject(tutorialRawAirTableFixture());
+
+      // then
+      expect(tuto).to.be.an.instanceof(Tutorial);
+      expect(tuto).to.deep.equal(expectedTuto);
+    });
+  });
+
   describe('#findByRecordIds', () => {
 
     it('should return an array of airtable tutorial data objects', function() {
       // given
       const rawTutorial1 = tutorialRawAirTableFixture();
-      rawTutorial1.id = 'FAKE_REC_ID_RAW_TUTORIAL_1' ;
+      rawTutorial1.id = 'FAKE_REC_ID_RAW_TUTORIAL_1';
 
       const rawTutorial2 = tutorialRawAirTableFixture();
-      rawTutorial2.id = 'FAKE_REC_ID_RAW_TUTORIAL_2' ;
+      rawTutorial2.id = 'FAKE_REC_ID_RAW_TUTORIAL_2';
 
       const rawTutorial3 = tutorialRawAirTableFixture();
-      rawTutorial3.id = 'FAKE_REC_ID_RAW_TUTORIAL_3' ;
+      rawTutorial3.id = 'FAKE_REC_ID_RAW_TUTORIAL_3';
 
       const records = [rawTutorial1, rawTutorial2, rawTutorial3];
       sinon.stub(airtable, 'findRecords').callsFake(makeAirtableFake(records));
@@ -51,7 +66,7 @@ describe('Unit | Infrastructure | Datasource | Airtable | TutorialDatasource', (
       });
     });
   });
-  
+
   describe('#get', () => {
 
     it('should call airtable on Tutoriels table with the id and return a datamodel Tutorial object', () => {
